+++ extensions/convert/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Convert Changelog
 
+## [Update] - {PR_MERGE_DATE}
+
+- Added support for alpha as percentage in hsl()
+- Fixed HEXA alpha
+- Extract hsl regex matching function and add unit tests
+- Bump `@raycast/eslint-config` and move it to devDependencies
+
 ## [Update] - 2026-06-25
 
 - Added support for modern space-separated hsl() color syntax